# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e2456320cfc6233ae9336d9
  title: Manage_Film_Example
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /film/:
    post:
      x-skaffolder-id: 5e2456340cfc6233ae933723
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Film
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                genre:
                  type: string
                title:
                  type: string
                year:
                  type: number
    get:
      x-skaffolder-id: 5e2456340cfc6233ae933727
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Film
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
  /film/{id}:
    delete:
      x-skaffolder-id: 5e2456340cfc6233ae933726
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e2456340cfc6233ae933720
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
    post:
      x-skaffolder-id: 5e2456340cfc6233ae933722
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Film
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
  /film/findBycast/{key}:
    get:
      x-skaffolder-id: 5e2456330cfc6233ae93370d
      x-skaffolder-name: findBycast
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: findBycast
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycast
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycast/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycast
      tags:
        - Film
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa cast da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
  /film/findByfilmMaker/{key}:
    get:
      x-skaffolder-id: 5e2456340cfc6233ae933719
      x-skaffolder-name: findByfilmMaker
      x-skaffolder-id-resource: 5e2456330cfc6233ae933705
      x-skaffolder-resource: Film
      x-skaffolder-crudAction: findByfilmMaker
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByfilmMaker
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByfilmMaker/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByfilmMaker
      tags:
        - Film
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa filmMaker da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
    
  #-----  FilmMaker  -----
...
  /electrobergsolutions/:
    post:
      x-skaffolder-id: 5e2457410cfc6233ae933792
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2457090cfc6233ae93378b
      x-skaffolder-resource: electrobergsolutions
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - electrobergsolutions
      parameters:
        - name: obj
          x-skaffolder-type: electrobergsolutions
          in: path
          description: Object to insert
          required: true
          schema:
            type: electrobergsolutions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/electrobergsolutions"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
    get:
      x-skaffolder-id: 5e2457380cfc6233ae93378f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2457090cfc6233ae93378b
      x-skaffolder-resource: electrobergsolutions
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF electrobergsolutions
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - electrobergsolutions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/electrobergsolutions"
  /electrobergsolutions/{id}:
    delete:
      x-skaffolder-id: 5e2457450cfc6233ae933794
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e2457090cfc6233ae93378b
      x-skaffolder-resource: electrobergsolutions
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - electrobergsolutions
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id electrobergsolutions
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e2457310cfc6233ae93378d
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e2457090cfc6233ae93378b
      x-skaffolder-resource: electrobergsolutions
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: electrobergsolutions
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - electrobergsolutions
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id electrobergsolutions
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/electrobergsolutions"
    post:
      x-skaffolder-id: 5e24573d0cfc6233ae933790
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e2457090cfc6233ae93378b
      x-skaffolder-resource: electrobergsolutions
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: electrobergsolutions
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - electrobergsolutions
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id electrobergsolutions
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/electrobergsolutions"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/electrobergsolutions"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Film:
      x-skaffolder-id: 5e2456330cfc6233ae933705
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2456330cfc6233ae9336fb
      x-skaffolder-url: /film
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        genre:
          type: string
          x-skaffolder-id-attr: 5e2456330cfc6233ae9336fe
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - Action
            - Crime
            - Fantasy
            - Horror
        title:
          type: string
          x-skaffolder-id-attr: 5e2456330cfc6233ae9336fc
          x-skaffolder-type: String
          x-skaffolder-required: true
        year:
          type: number
          x-skaffolder-id-attr: 5e2456330cfc6233ae9336fd
          x-skaffolder-type: Number
      x-skaffolder-relations: 
        cast:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e2456330cfc6233ae9336fb
          x-skaffolder-ent2: 5e2456320cfc6233ae9336f0
        filmMaker:
          x-skaffolder-type: 1:m
          x-skaffolder-required: true
          x-skaffolder-ent1: 5e2456330cfc6233ae9336fb
          x-skaffolder-ent2: 5e2456330cfc6233ae9336f6
      required:
        - _id
        - title
...
    electrobergsolutions:
      x-skaffolder-id: 5e2457090cfc6233ae93378b
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2457090cfc6233ae93378a
      x-skaffolder-url: /electrobergsolutions
      properties:
        _id:
          type: string
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
      
    
    
  # Databases
...