# src/main/java/com/manage_film_example/controller/base/ElectrobergsolutionsBaseController.java
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5e2456320cfc6233ae9336d8
*
* You will get 10% discount for each one of your friends
* 
*/
package com.manage_film_example.controller.base;

import org.springframework.security.access.annotation.Secured;
import java.util.ArrayList;
import java.util.List;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import com.manage_film_example.db.manage_film_example_db.service.ElectrobergsolutionsService;
import com.manage_film_example.db.manage_film_example_db.mapper.Electrobergsolutions;

public class ElectrobergsolutionsBaseController {

	ElectrobergsolutionsService electrobergsolutionsService = new ElectrobergsolutionsService();



//CRUD METHODS


    //CRUD - CREATE
    @Secured({"ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/electrobergsolutions", method = RequestMethod.POST, headers = "Accept=application/json")
	public Electrobergsolutions insert(@RequestBody Electrobergsolutions obj) {
		Electrobergsolutions result = electrobergsolutionsService.insert(obj);

	    
		
		return result;
	}

	
    //CRUD - REMOVE
    @Secured({"ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/electrobergsolutions/{id}", method = RequestMethod.DELETE, headers = "Accept=application/json")
	public void delete(@PathVariable("id") Long id) {
		electrobergsolutionsService.delete(id);
	}
	
	
    //CRUD - GET ONE
    @Secured({"ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/electrobergsolutions/{id}", method = RequestMethod.GET, headers = "Accept=application/json")
	public Electrobergsolutions get(@PathVariable Long id) {
		Electrobergsolutions obj = electrobergsolutionsService.get(id);
		
		
		
		return obj;
	}
	
	
    //CRUD - GET LIST
    @Secured({"ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/electrobergsolutions", method = RequestMethod.GET, headers = "Accept=application/json")
	public List<Electrobergsolutions> getList() {
		List<Electrobergsolutions> list = electrobergsolutionsService.getList();
		return list;
	}
	
	

    //CRUD - EDIT
    @Secured({"ROLE_PRIVATE_USER" })
	@RequestMapping(value = "/electrobergsolutions/{id}", method = RequestMethod.POST, headers = "Accept=application/json")
	public Electrobergsolutions update(@RequestBody Electrobergsolutions obj, @PathVariable("id") Long id) {
		Electrobergsolutions result = electrobergsolutionsService.update(obj, id);

	    
		
		return result;
	}
	


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */


	
}